refactor: split analysis audio module
